pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "jumashafara",email = "jumashafara0@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=6.0,<7.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-cors-headers (>=4.6.0,<5.0.0)",
    "django-filter (>=24.3,<25.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
